Decompile func00029ffc

+++ src/include/lib/lib_24e40.h
@@ -57,7 +57,7 @@ s32 func00028914(struct tile *start, struct tile *end, struct coord *pos, f32 wi
 void func00028df0(struct coord *pos, f32 width, s16 *rooms, u32 types, u16 arg4, u32 arg5, f32 ymax, f32 ymin, struct collisionthing *arg8, s32 arg9);
 void func0002901c(struct coord *pos, struct coord *dist, f32 width, struct collisionthing *arg3);
 f32 func000296a0(struct collisionthing *arg0, struct coord *pos, struct tilething **arg2, f32 width);
-bool func00029ffc(struct coord *pos, f32 width, f32 foreheadheight, f32 inversefeettoeyesheight, s16 *rooms, u32 arg5, struct coord *laddernormal);
+bool func00029ffc(struct coord *pos, f32 width, f32 foreheadheight, f32 inversefeettoeyesheight, s16 *rooms, u16 arg5, struct coord *laddernormal);
 u32 func0002a13c(void);
 f32 cdFindGroundY(struct coord *pos, f32 width, s16 *rooms, u16 *floorcol, u8 *floortype, u16 *floorflags, s16 *floorroom, s32 *inlift, struct prop **lift);
 f32 func0002a324(void);
